+++ agent/src/index.test.ts
@@ -5,19 +5,42 @@ import os from 'os'
 import path from 'path'
 
 import { afterAll, beforeAll, describe, expect, it } from 'vitest'
+import * as vscode from 'vscode'
 import { Uri } from 'vscode'
 
-import { type ChatMessage } from '@sourcegraph/cody-shared'
+import { type ChatMessage, type ContextFile } from '@sourcegraph/cody-shared'
 
 import type { ExtensionMessage } from '../../vscode/src/chat/protocol'
 
 import { AgentTextDocument } from './AgentTextDocument'
 import { MessageHandler } from './jsonrpc-alias'
-import { type ClientInfo, type ServerInfo } from './protocol-alias'
+import { type ClientInfo, type ProgressReportParams, type ProgressStartParams, type ServerInfo } from './protocol-alias'
+
+type ProgressMessage = ProgressStartMessage | ProgressReportMessage | ProgressEndMessage
+interface ProgressStartMessage {
+    method: 'progress/start'
+    id: string
+    message: ProgressStartParams
+}
+interface ProgressReportMessage {
+    method: 'progress/report'
+    id: string
+    message: ProgressReportParams
+}
+interface ProgressEndMessage {
+    method: 'progress/end'
+    id: string
+    message: {}
+}
 
 export class TestClient extends MessageHandler {
     public info: ClientInfo
     public agentProcess?: ChildProcessWithoutNullStreams
+    // Array of all raw `progress/*` notification. Typed as `any` because
+    // start/end/report have different types.
+    public progressMessages: ProgressMessage[] = []
+    public progressIDs = new Map<string, number>()
+    public progressStartEvents = new vscode.EventEmitter<ProgressStartParams>()
 
     constructor(
         public readonly name: string,
@@ -28,10 +51,32 @@ export class TestClient extends MessageHandler {
         this.name = name
         this.info = this.getClientInfo()
 
+        this.registerNotification('progress/start', message => {
+            this.progressStartEvents.fire(message)
+            message.id = this.progressID(message.id)
+            this.progressMessages.push({ method: 'progress/start', id: message.id, message })
+        })
+        this.registerNotification('progress/report', message => {
+            message.id = this.progressID(message.id)
+            this.progressMessages.push({ method: 'progress/report', id: message.id, message })
+        })
+        this.registerNotification('progress/end', ({ id }) => {
+            this.progressMessages.push({ method: 'progress/end', id: this.progressID(id), message: {} })
+        })
         this.registerNotification('debug/message', message => {
             // Uncomment below to see `logDebug` messages.
             // console.log(`${message.channel}: ${message.message}`)
         })
+    }
+
+    private progressID(id: string): string {
+        const fromCache = this.progressIDs.get(id)
+        if (fromCache !== undefined) {
+            return `ID_${fromCache}`
+        }
+        const freshID = this.progressIDs.size
+        this.progressIDs.set(id, freshID)
+        return `ID_${freshID}`
     }
 
     public async initialize() {
@@ -60,7 +105,10 @@ export class TestClient extends MessageHandler {
         }
     }
 
-    public async sendSingleMessage(text: string): Promise<ChatMessage | undefined> {
+    public async sendSingleMessage(
+        text: string,
+        params?: { addEnhancedContext?: boolean; contextFiles?: ContextFile[] }
+    ): Promise<ChatMessage | undefined> {
         const id = await this.request('chat/new', null)
         const reply = await this.request('chat/submitMessage', {
             id,
@@ -68,8 +116,8 @@ export class TestClient extends MessageHandler {
                 command: 'submit',
                 text,
                 submitType: 'user',
-                addEnhancedContext: false,
-                contextFiles: [],
+                addEnhancedContext: params?.addEnhancedContext ?? false,
+                contextFiles: params?.contextFiles,
             },
         })
 
@@ -122,7 +170,7 @@ export class TestClient extends MessageHandler {
         const recordingDirectory = path.join(agentDir, 'recordings')
         const agentScript = path.join(agentDir, 'dist', 'index.js')
 
-        return spawn('node', ['--inspect', '--enable-source-maps', agentScript, 'jsonrpc'], {
+        return spawn('node', ['--enable-source-maps', agentScript, 'jsonrpc'], {
             stdio: 'pipe',
             cwd: agentDir,
             env: {
@@ -145,6 +193,9 @@ export class TestClient extends MessageHandler {
             version: 'v1',
             workspaceRootUri: workspaceRootUri.toString(),
             workspaceRootPath: workspaceRootUri.fsPath,
+            capabilities: {
+                progressBars: 'enabled',
+            },
             extensionConfiguration: {
                 anonymousUserID: this.name + 'abcde1234',
                 accessToken: this.accessToken ?? 'sgp_RRRRRRRREEEEEEEDDDDDDAAACCCCCTEEEEEEEDDD',
@@ -162,6 +213,30 @@ export class TestClient extends MessageHandler {
     }
 }
 
+const explainPollyError = `
+
+    ===================================================[ NOTICE ]=======================================================
+    If you get PollyError or unexpected diff, you might need to update recordings to match your changes.
+    Run the following commands locally to update the recordings:
+
+      export SRC_ACCESS_TOKEN=YOUR_TOKEN
+      export SRC_ACCESS_TOKEN_WITH_RATE_LIMIT=RATE_LIMITED_TOKEN # see https://sourcegraph.slack.com/archives/C059N5FRYG3/p1702990080820699
+      export SRC_ENDPOINT=https://sourcegraph.com
+      pnpm update-agent-recordings
+      # Press 'u' to update the snapshots if the new behavior makes sense. It's
+      # normal that the LLM returns minor changes to the wording.
+      git commit -am "Update agent recordings"
+
+
+    More details in https://github.com/sourcegraph/cody/tree/main/agent#updating-the-polly-http-recordings
+    ====================================================================================================================
+
+    `
+
+const prototypePath = path.join(__dirname, '__tests__', 'example-ts')
+const workspaceRootUri = Uri.file(path.join(os.tmpdir(), 'cody-vscode-shim-test'))
+const workspaceRootPath = workspaceRootUri.fsPath
+
 describe('Agent', () => {
     // Uncomment the code block below to disable agent tests. Feel free to do this to unblock
     // merging a PR if the agent tests are failing. If you decide to uncomment this block, please
@@ -174,42 +249,31 @@ describe('Agent', () => {
 
     const dotcom = 'https://sourcegraph.com'
     if (process.env.CODY_RECORDING_MODE === 'record' || process.env.CODY_RECORD_IF_MISSING === 'true') {
-        console.log('Because CODY_RECORDING_MODE=record, validating that you are authenticated to sourcegraph.com')
         execSync('src login', { stdio: 'inherit' })
         assert.strictEqual(process.env.SRC_ENDPOINT, dotcom, 'SRC_ENDPOINT must be https://sourcegraph.com')
     }
-
-    const explainPollyError = `
-
-    ===================================================[ NOTICE ]=======================================================
-    If you get PollyError or unexpeced diff, you might need to update recordings to match your changes.
-    Please check https://github.com/sourcegraph/cody/tree/main/agent#updating-the-polly-http-recordings for the details.
-    ====================================================================================================================
-
-    `
 
     if (process.env.VITEST_ONLY && !process.env.VITEST_ONLY.includes('Agent')) {
         it('Agent tests are skipped due to VITEST_ONLY environment variable', () => {})
         return
     }
 
-    const prototypePath = path.join(__dirname, '__tests__', 'example-ts')
-    const workspaceRootUri = Uri.file(path.join(os.tmpdir(), 'cody-vscode-shim-test'))
-    const workspaceRootPath = workspaceRootUri.fsPath
-
     const client = new TestClient('defaultClient', process.env.SRC_ACCESS_TOKEN)
-    const rateLimitedClient = new TestClient('rateLimitedClient', process.env.SRC_ACCESS_TOKEN_WITH_RATE_LIMIT)
 
     // Bundle the agent. When running `pnpm run test`, vitest doesn't re-run this step.
-    execSync('pnpm run build', { cwd: client.getAgentDir(), stdio: 'inherit' })
+    //
+    // ⚠️ If this line fails when running unit tests, chances are that the error is being swallowed.
+    // To see the full error, run this file in isolation:
+    //
+    //   pnpm test agent/src/index.test.ts
+    execSync('pnpm run build:agent', { cwd: client.getAgentDir(), stdio: 'inherit' })
 
     // Initialize inside beforeAll so that subsequent tests are skipped if initialization fails.
     beforeAll(async () => {
         await fspromises.mkdir(workspaceRootPath, { recursive: true })
         await fspromises.cp(prototypePath, workspaceRootPath, { recursive: true })
         await client.initialize()
-        await rateLimitedClient.initialize()
-    }, 1000000)
+    }, 10_000)
 
     it('handles config changes correctly', () => {
         // Send two config change notifications because this is what the
@@ -292,15 +356,14 @@ describe('Agent', () => {
     }, 10_000)
 
     it('allows us to send a very short chat message', async () => {
-        await openFile(animalUri)
         const lastMessage = await client.sendSingleMessage('Hello!')
         expect(lastMessage).toMatchInlineSnapshot(
             `
           {
             "contextFiles": [],
-            "displayText": " Hello! Nice to meet you.",
+            "displayText": " Hello there!",
             "speaker": "assistant",
-            "text": " Hello! Nice to meet you.",
+            "text": " Hello there!",
           }
         `,
             explainPollyError
@@ -308,12 +371,8 @@ describe('Agent', () => {
     }, 20_000)
 
     it('allows us to send a longer chat message', async () => {
-        await openFile(animalUri)
         const lastMessage = await client.sendSingleMessage('Generate simple hello world function in java!')
-        const trimmedMessage = lastMessage?.text
-            ?.split('\n')
-            .map(line => line.trimEnd())
-            .join('\n')
+        const trimmedMessage = trimEndOfLine(lastMessage?.text ?? '')
         expect(trimmedMessage).toMatchInlineSnapshot(
             `
           " Here is a simple Hello World program in Java:
@@ -328,25 +387,65 @@ describe('Agent', () => {
 
           To break this down:
 
-          - The code is wrapped in a class called Main. In Java, code must be inside a class.
+          - The code is wrapped in a class called Main. In Java, code must be contained within classes.
 
           - The main method is the entry point of the program. It is marked as static so it can be run without creating an instance of Main.
 
           - The main method accepts a String array called args as a parameter. This contains any command line arguments passed to the program.
 
-          - Inside the main method, we call System.out.println(\\"Hello World\\"); to print the text \\"Hello World!\\" to the console.
+          - System.out.println prints the text \\"Hello World!\\" to the console.
 
-          - The println method prints the text and a newline character.
+          To run this:
 
-          So in summary, this program defines a Main class with a static main method that prints \\"Hello World!\\" when executed. This is the simplest Hello World program in Java."
+          1. Save the code in a file called Main.java
+          2. Compile it with: javac Main.java
+          3. Run it with: java Main
+
+          This will print \\"Hello World!\\" to the console when executed."
         `,
             explainPollyError
         )
     }, 20_000)
 
-    it('get rate limit error if exceeding usage on rate limited account', async () => {
-        const lastMessage = await rateLimitedClient.sendSingleMessage('sqrt(9)')
-        expect(lastMessage?.error?.name).toMatchInlineSnapshot('"RateLimitError"', explainPollyError)
+    // This test is skipped because it shells out to `symf expand-query`, which
+    // requires an access token to send an llm request and is, therefore, not
+    // able to return stable results in replay mode. Also, we don't have an
+    // access token in ci so this test can only pass when running locally (for
+    // now).
+    it('allows us to send a chat message with enhanced context enabled', async () => {
+        await openFile(animalUri)
+        await client.request('command/execute', { command: 'cody.search.index-update' })
+        const lastMessage = await client.sendSingleMessage(
+            'Write a class Dog that implements the Animal interface in my workspace. Only show the code, no explanation needed.',
+            {
+                addEnhancedContext: true,
+            }
+        )
+        // TODO: make this test return a TypeScript implementation of
+        // `animal.ts`. It currently doesn't do this because the workspace root
+        // is not a git directory and symf reports some git-related error.
+        expect(trimEndOfLine(lastMessage?.text ?? '')).toMatchInlineSnapshot(
+            `
+          " Here is the code for the Dog class implementing the Animal interface:
+
+          \`\`\`java
+          public class Dog implements Animal {
+
+            @Override
+            public void makeSound() {
+              System.out.println(\\"Woof!\\");
+            }
+
+            @Override
+            public void move() {
+              System.out.println(\\"The dog runs.\\");
+            }
+
+          }
+          \`\`\`"
+        `,
+            explainPollyError
+        )
     }, 20_000)
 
     // TODO Fix test - fails intermittently on CI
@@ -356,10 +455,107 @@ describe('Agent', () => {
         await client.request('recipes/execute', { id: 'chat-question', humanChatInput: 'How do I implement sum?' })
     }, 600)
 
+    describe('progress bars', () => {
+        it('messages are sent', async () => {
+            const { result } = await client.request('testing/progress', { title: 'Susan' })
+            expect(result).toStrictEqual('Hello Susan')
+            let progressID: string | undefined
+            for (const message of client.progressMessages) {
+                if (message.method === 'progress/start' && message.message.options.title === 'testing/progress') {
+                    progressID = message.message.id
+                    break
+                }
+            }
+            assert(progressID !== undefined, JSON.stringify(client.progressMessages))
+            const messages = client.progressMessages
+                .filter(message => message.id === progressID)
+                .map(({ method, message }) => [method, { ...message, id: 'THE_ID' }])
+            expect(messages).toMatchInlineSnapshot(`
+              [
+                [
+                  "progress/start",
+                  {
+                    "id": "THE_ID",
+                    "options": {
+                      "cancellable": true,
+                      "location": "Notification",
+                      "title": "testing/progress",
+                    },
+                  },
+                ],
+                [
+                  "progress/report",
+                  {
+                    "id": "THE_ID",
+                    "message": "message1",
+                  },
+                ],
+                [
+                  "progress/report",
+                  {
+                    "id": "THE_ID",
+                    "increment": 50,
+                  },
+                ],
+                [
+                  "progress/report",
+                  {
+                    "id": "THE_ID",
+                    "increment": 50,
+                  },
+                ],
+                [
+                  "progress/end",
+                  {
+                    "id": "THE_ID",
+                  },
+                ],
+              ]
+            `)
+        })
+        it('progress can be cancelled', async () => {
+            const disposable = client.progressStartEvents.event(params => {
+                if (params.options.title === 'testing/progressCancelation') {
+                    client.notify('progress/cancel', { id: params.id })
+                }
+            })
+            try {
+                const { result } = await client.request('testing/progressCancelation', { title: 'Leona' })
+                expect(result).toStrictEqual("request with title 'Leona' cancelled")
+            } finally {
+                disposable.dispose()
+            }
+        })
+    })
+
+    describe('RateLimitedAgent', () => {
+        const rateLimitedClient = new TestClient('rateLimitedClient', process.env.SRC_ACCESS_TOKEN_WITH_RATE_LIMIT)
+        // Initialize inside beforeAll so that subsequent tests are skipped if initialization fails.
+        beforeAll(async () => {
+            await rateLimitedClient.initialize()
+        }, 10_000)
+
+        it('get rate limit error if exceeding usage on rate limited account', async () => {
+            const lastMessage = await rateLimitedClient.sendSingleMessage('sqrt(9)')
+            expect(lastMessage?.error?.name).toMatchInlineSnapshot('"RateLimitError"', explainPollyError)
+        }, 20_000)
+
+        afterAll(async () => {
+            await rateLimitedClient.shutdownAndExit()
+            // Long timeout because to allow Polly.js to persist HTTP recordings
+        }, 20_000)
+    })
+
     afterAll(async () => {
         await fspromises.rm(workspaceRootPath, { recursive: true, force: true })
         await client.shutdownAndExit()
-        await rateLimitedClient.shutdownAndExit()
         // Long timeout because to allow Polly.js to persist HTTP recordings
     }, 20_000)
 })
+
+function trimEndOfLine(text: string): string {
+    return text
+        .split('\n')
+        .map(line => line.trimEnd())
+        .join('\n')
+}