+++ vscode/src/configuration.ts
@@ -102,7 +102,7 @@ export function getConfiguration(config: ConfigGetter = vscode.workspace.getConf
         autocompleteExperimentalGraphContext,
         experimentalChatPredictions: getHiddenSetting('experimental.chatPredictions', isTesting),
         experimentalSimpleChatContext: getHiddenSetting('experimental.simpleChatContext', true),
-        experimentalSymfContext: getHiddenSetting('experimental.symfContext', false),
+        experimentalSymfContext: getHiddenSetting('experimental.symfContext', true),
 
         experimentalGuardrails: getHiddenSetting('experimental.guardrails', isTesting),
         experimentalLocalSymbols: getHiddenSetting('experimental.localSymbols', false),
@@ -114,6 +114,10 @@ export function getConfiguration(config: ConfigGetter = vscode.workspace.getConf
         ),
 
         autocompleteExperimentalHotStreak: getHiddenSetting('autocomplete.experimental.hotStreak', false),
+        autocompleteExperimentalOllamaOptions: getHiddenSetting('autocomplete.experimental.ollamaOptions', {
+            url: 'http://localhost:11434',
+            model: 'codellama:7b-code',
+        }),
 
         // Note: In spirit, we try to minimize agent-specific code paths in the VSC extension.
         // We currently use this flag for the agent to provide more helpful error messages